feat: timeSent를 적절한 객체가 가지도록 수정
